safari-534.53-branch/LayoutTests/storage/indexeddb/mozilla/key-requirements-delete-null-key-expected.txt

Test IndexedDB's behavior deleting entry with no key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;
PASS indexedDB == null is false
indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('version 1')
Deleted all object stores.
objectStore = db.createObjectStore('bar');
Expecting exception from objectStore.delete(null);
PASS Exception was thrown.
PASS code is webkitIDBDatabaseException.DATA_ERR
PASS successfullyParsed is true

TEST COMPLETE
